Extract FlexibleDataModelsTestBase
