+++ src/test/scala/cognite/spark/v1/FlexibleDataModelConnectionRelationTest.scala
@@ -2,14 +2,13 @@ package cognite.spark.v1
 
 import org.scalatest.{FlatSpec, Matchers}
 
-class FlexibleDataModelConnectionRelationTest extends FlatSpec with Matchers with SparkTest {
-//  private val clientId = sys.env("TEST_CLIENT_ID_BLUEFIELD")
-//  private val clientSecret = sys.env("TEST_CLIENT_SECRET_BLUEFIELD")
-//  private val aadTenant = sys.env("TEST_AAD_TENANT_BLUEFIELD")
-//  private val tokenUri = s"https://login.microsoftonline.com/$aadTenant/oauth2/v2.0/token"
-//  private val client = getBlufieldClient()
-//
-//  private val spaceExternalId = "testSpaceForSparkDs"
+class FlexibleDataModelConnectionRelationTest
+    extends FlatSpec
+    with Matchers
+    with SparkTest
+    with FlexibleDataModelsTestBase {
+
+  //  client.spacesv3.createItems(Seq(SpaceCreateDefinition(spaceExternalId))).unsafeRunSync()
 
   it should "pass" in {
     1 shouldBe 1
